+++ CHANGELOG.md
@@ -7,6 +7,7 @@
     - Dataset fromIterator method accepts any iterable
     - Column Picker throws exception if column not found
     - Better hyper-parameter stringification
+    - Improved Dataset exception messages
 
 - 0.0.19-beta
     - Added SiLU self-stabilizing neural network activation function